+++ src/question.ts
@@ -58,6 +58,7 @@ export class Question extends SurveyElement
   customWidgetData = { isNeedRender: true };
   focusCallback: () => void;
   surveyLoadCallback: () => void;
+  displayValueCallback: (text: string) => string;
 
   private textPreProcessor: TextPreProcessor;
   private conditionEnabelRunner: ConditionRunner;
@@ -281,7 +282,7 @@ export class Question extends SurveyElement
    * @see titleLocation
    */
   public get hideNumber(): boolean {
-    return this.getPropertyValue("hideNumber", false);
+    return this.getPropertyValue("hideNumber");
   }
   public set hideNumber(val: boolean) {
     this.setPropertyValue("hideNumber", val);
@@ -298,6 +299,9 @@ export class Question extends SurveyElement
    */
   public getType(): string {
     return "question";
+  }
+  public get isQuestion(): boolean {
+    return true;
   }
   /**
    * Move question to a new container Page/Panel. Add as a last element if insertBefore parameter is not used or inserted into the given index,
@@ -1107,6 +1111,7 @@ export class Question extends SurveyElement
   }
   protected onSetData() {
     super.onSetData();
+    if (!this.survey) return;
     this.initDataFromSurvey();
     this.onSurveyValueChanged(this.value);
     this.updateValueWithDefaults();
@@ -1164,8 +1169,10 @@ export class Question extends SurveyElement
    * Clear the question value. It clears the question comment as well.
    */
   public clearValue() {
-    this.value = null;
-    this.comment = null;
+    if (this.value !== undefined) {
+      this.value = undefined;
+    }
+    this.comment = undefined;
   }
   public unbindValue() {
     this.clearValue();
@@ -1212,16 +1219,23 @@ export class Question extends SurveyElement
    * @param value use this parameter, if you want to get display value for this value and not question.value. It is undefined by default.
    */
   public getDisplayValue(keysAsText: boolean, value: any = undefined): any {
+    var res = this.calcDisplayValue(keysAsText, value);
+    return !!this.displayValueCallback ? this.displayValueCallback(res) : res;
+  }
+  private calcDisplayValue(keysAsText: boolean, value: any = undefined): any {
     if (this.customWidget) {
       var res = this.customWidget.getDisplayValue(this, value);
       if (res) return res;
     }
     value = value == undefined ? this.createValueCopy() : value;
-    if (this.isValueEmpty(value)) return "";
+    if (this.isValueEmpty(value)) return this.getDisplayValueEmpty();
     return this.getDisplayValueCore(keysAsText, value);
   }
   protected getDisplayValueCore(keyAsText: boolean, value: any): any {
     return value;
+  }
+  protected getDisplayValueEmpty(): string {
+    return "";
   }
   /**
    * Set the default value to the question. It will be assign to the question on loading the survey from JSON or adding a question to the survey or on setting this property of the value is empty.
@@ -1550,10 +1564,13 @@ export class Question extends SurveyElement
   }
   private checkForErrors(isOnValueChanged: boolean): Array<SurveyError> {
     var qErrors = new Array<SurveyError>();
-    if (this.isVisible && !this.isReadOnly) {
+    if (this.isVisible && this.canCollectErrors()) {
       this.collectErrors(qErrors, isOnValueChanged);
     }
     return qErrors;
+  }
+  protected canCollectErrors(): boolean {
+    return !this.isReadOnly;
   }
   private collectErrors(
     qErrors: Array<SurveyError>,