+++ tests/jsonobjecttests.ts
@@ -2409,3 +2409,22 @@ QUnit.test("itemvalue enableIf property visibility test", function(assert) {
     "We show enableIf for all other properties"
   );
 });
+QUnit.test("Change default value for question.hideNumber", function(assert) {
+  assert.equal(
+    new Question("q1").hideNumber,
+    false,
+    "By default hideNumber returns false"
+  );
+  Serializer.findProperty("question", "hideNumber").defaultValue = true;
+  assert.equal(
+    new Question("q1").hideNumber,
+    true,
+    "We have override hide number"
+  );
+  Serializer.findProperty("question", "hideNumber").defaultValue = undefined;
+  assert.equal(
+    new Question("q1").hideNumber,
+    false,
+    "We made hideNumber false by default again"
+  );
+});