+++ tests/surveytests.ts
@@ -13440,3 +13440,135 @@ QUnit.test(
     SurveyElement.FocusElement = oldFunc;
   }
 );
+QUnit.test(
+  "Test SurveyElement isPage, isPanel and isQuestion properties",
+  function(assert) {
+    var survey = new SurveyModel({
+      elements: [
+        {
+          type: "panel",
+          name: "panel",
+          elements: [
+            {
+              type: "text",
+              name: "question",
+            },
+          ],
+        },
+      ],
+    });
+    var page = survey.currentPage;
+    var panel = <PanelModel>survey.getPanelByName("panel");
+    var question = survey.getQuestionByName("question");
+    assert.equal(page.isPage, true, "Page is page");
+    assert.equal(page.isPanel, false, "Page is not panel");
+    assert.equal(page.isQuestion, false, "Page is not question");
+    assert.equal(panel.isPage, false, "Panel is not page");
+    assert.equal(panel.isPanel, true, "Panel is panel");
+    assert.equal(panel.isQuestion, false, "Panel is not question");
+    assert.equal(question.isPage, false, "Question is not page");
+    assert.equal(question.isPanel, false, "Question is not panel");
+    assert.equal(question.isQuestion, true, "Question is question");
+  }
+);
+QUnit.test("Test survey renderedHasTitle/renderedHasLogo properties", function(
+  assert
+) {
+  var survey = new SurveyModel();
+  assert.equal(
+    survey.renderedHasHeader,
+    false,
+    "hasHeader, title and logo are invisible"
+  );
+  assert.equal(survey.renderedHasTitle, false, "There is not title");
+  survey.title = "title";
+  assert.equal(survey.renderedHasTitle, true, "There is title");
+  assert.equal(survey.renderedHasHeader, true, "hasHeader, title is visible");
+  survey.showTitle = false;
+  assert.equal(survey.renderedHasTitle, false, "hasTitle is false");
+
+  assert.equal(survey.renderedHasLogo, false, "There is not logo");
+  survey.logo = "logo";
+  assert.equal(survey.renderedHasLogo, true, "There is logo");
+  survey.logoPosition = "none";
+  assert.equal(survey.renderedHasTitle, false, "logo position is 'none'");
+
+  survey.setDesignMode(true);
+  assert.equal(survey.renderedHasTitle, true, "There is title, design");
+  assert.equal(survey.renderedHasLogo, true, "There is logo, design");
+  assert.equal(
+    survey.isLogoBefore,
+    false,
+    "We do not render logo before at design, design"
+  );
+  assert.equal(
+    survey.isLogoAfter,
+    true,
+    "We do render logo after at design, design"
+  );
+  assert.equal(
+    survey.renderedHasHeader,
+    true,
+    "hasHeader, properties are visible"
+  );
+
+  Serializer.findProperty("survey", "title").visible = false;
+  Serializer.findProperty("survey", "logo").visible = false;
+  assert.equal(
+    survey.renderedHasTitle,
+    false,
+    "There is no title, design, property invisible"
+  );
+  assert.equal(
+    survey.renderedHasLogo,
+    false,
+    "There is no logo, design, property invisible"
+  );
+  assert.equal(
+    survey.isLogoAfter,
+    false,
+    "We do not render logo after since the property is hidden, design"
+  );
+  assert.equal(
+    survey.renderedHasHeader,
+    false,
+    "hasHeader, properties are invisible"
+  );
+  Serializer.findProperty("survey", "title").visible = true;
+  Serializer.findProperty("survey", "logo").visible = true;
+});
+QUnit.test("Test survey renderedHasTitle/renderedHasLogo properties", function(
+  assert
+) {
+  var survey = new SurveyModel();
+  var trigger = new SurveyTriggerSetValue();
+  survey.triggers.push(trigger);
+
+  var surveyPropertyName;
+  var triggerPropertyName;
+  survey.onPropertyValueChangedCallback = (
+    name: string,
+    oldValue: any,
+    newValue: any,
+    sender: Base,
+    arrayChanges: any
+  ) => {
+    surveyPropertyName = name;
+  };
+  trigger.onPropertyValueChangedCallback = (
+    name: string,
+    oldValue: any,
+    newValue: any,
+    sender: Base,
+    arrayChanges: any
+  ) => {
+    triggerPropertyName = name;
+  };
+  trigger.setValue = "test";
+  assert.equal(surveyPropertyName, "setValue", "get notification from survey");
+  assert.equal(
+    triggerPropertyName,
+    "setValue",
+    "get notification from trigger"
+  );
+});